requirements.txt
fastapi==0.74.1
Pillow==9.1.0
pydantic==1.9.0
requests==2.27.1
rich==11.2.0
uvicorn==0.17.5